Isolate notification logic - working
