+++ app/createRide.tsx
@@ -15,9 +15,10 @@ import React, { useState, useEffect, useRef } from "react";
 import { ScrollView, StyleSheet, Switch, Text, View, Button, Platform } from "react-native";
 import Title from "../components/Title";
 import { RideDataType } from "./rsvp";
-import * as Device from 'expo-device';
-import * as Notifications from 'expo-notifications';
-import Constants from 'expo-constants';
+import {
+  registerForPushNotificationsAsync,
+  scheduleRideNotification
+} from "@/utils/notifications";
 
 export type AllowedGenders = "Co-ed" | "Female" | "Male";
 type RecurrenceFrequency = "daily" | "weekly" | "monthly";
@@ -27,88 +28,6 @@ type RideFormData = Omit<RideDataType, "id" | "departs" | "returns"> & {
   recurrenceFrequency: RecurrenceFrequency;
   numOccurrences: string;
 };
-
-//////////////////// Notification Logic Functions //////////////////////////////////////
-Notifications.setNotificationHandler({
-  handleNotification: async () => ({
-    shouldPlaySound: true,
-    shouldSetBadge: true,
-    shouldShowBanner: true,
-    shouldShowList: true,
-  }),
-});
-
-function handleRegistrationError(errorMessage: string) {
-  alert(errorMessage);
-  throw new Error(errorMessage);
-}
-
-async function registerForPushNotificationsAsync() {
-  if (Platform.OS === 'android') {
-    await Notifications.setNotificationChannelAsync('default', {
-      name: 'default',
-      importance: Notifications.AndroidImportance.MAX,
-      vibrationPattern: [0, 250, 250, 250],
-      lightColor: '#FF231F7C',
-    });
-  }
-
-  if (Device.isDevice) {
-    const { status: existingStatus } = await Notifications.getPermissionsAsync();
-    let finalStatus = existingStatus;
-    if (existingStatus !== 'granted') {
-      const { status } = await Notifications.requestPermissionsAsync();
-      finalStatus = status;
-    }
-    if (finalStatus !== 'granted') {
-      handleRegistrationError('Permission not granted to get push token for push notification!');
-      return;
-    }
-    const projectId =
-      Constants?.expoConfig?.extra?.eas?.projectId ?? Constants?.easConfig?.projectId;
-    if (!projectId) {
-      handleRegistrationError('Project ID not found');
-    }
-    try {
-      const pushTokenString = (
-        await Notifications.getExpoPushTokenAsync({
-          projectId,
-        })
-      ).data;
-      console.log(pushTokenString);
-      return pushTokenString;
-    } catch (e: unknown) {
-      handleRegistrationError(`${e}`);
-    }
-  } else {
-    handleRegistrationError('Must use physical device for push notifications');
-  }
-}
-
-async function scheduleRideNotification(departsDate: Date) {
-  // 10 minutes before
-  const triggerTime = new Date(departsDate.getTime() - 10 * 60 * 1000);
-
-  // If the trigger time is already passed, skip scheduling
-  if (triggerTime <= new Date()) {
-    console.log("Skipping notification: departure is too soon or time already passed");
-    return;
-  }
-
-  await Notifications.scheduleNotificationAsync({
-    content: {
-      title: "Upcoming Ride Reminder 🚗",
-      body: "Your ride is coming up in 10 minutes!",
-      sound: "default",
-      data: { type: "ride-reminder" },
-    },
-    trigger: triggerTime,
-  });
-
-  console.log("Notification scheduled for:", triggerTime.toString());
-}
-
-//////////////////// End of Notification Logic Functions //////////////////////////////////////
 
 export default function CreateRide() {
   // Local state for date and time inputs (to prevent overriding each other)
